requirements.txt
appdirs==1.4.4
asgiref==3.8.1
ayon-python-api==1.0.12
certifi==2025.1.31
charset-normalizer==3.4.1
Django==4.2.20
django-auth-ldap==5.1.0
django-filter==25.1
django-tables2==2.7.5
dnspython==2.7.0
idna==3.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pyldap==3.0.0.post1
pymongo==4.11.2
python-dotenv==1.0.1
python-ldap==3.4.4
requests==2.32.3
sqlparse==0.5.3
typing_extensions==4.12.2
Unidecode==1.3.8
urllib3==2.3.0
